refactor: rename hostresolver package to host_resolver
