+++ src/infrastructure/network/host_resolver/dial_resolver_test.go
@@ -1,4 +1,4 @@
-package hostresolver
+package host_resolver
 
 import (
 	"errors"
